Calculate the correct diff root directory
Limited to CVS diffs for now.

+++ patch.sh
@@ -1,4 +1,9 @@
+. "${EXECDIR}/util.sh"
+
 [ -z "$DIFF" ] && exit 0
 
-cp "$DIFF" "${LOGDIR}/src.diff"
-(cd "$BSDSRCDIR" && patch -Es) <"$DIFF"
+[ -e "${LOGDIR}/src.diff" ] || cp "$DIFF" "${LOGDIR}/src.diff"
+
+SRCDIR="$(diff_root "$DIFF")"
+[ -z "$SRCDIR" ] && SRCDIR="$BSDSRCDIR"
+(cd "$SRCDIR" && patch -Es) <"$DIFF"
